Made the time formater into a reusable component
